[feat]Refactored all routes into one file
